Create list-pocket-residues.py
This script is very very very veeeery crude! It should work only for zinc enzymes, but it's trivial to see a way to improve it to other metalloenzymes as well as add some routines to verify if the user is providing the right thing as the first argument. In near future I will add these changes, but feel free to make your own. Cheers :D